implemented settings command

+++ Application/Extensions/RandomExtensions.cs
@@ -1,10 +1,10 @@
 namespace Application.Extensions;
 
 internal static class RandomExtensions {
-    public static bool WithChance(this Random randomizer, double chance) {
-        if (chance is < 0.0 or > 1.0)
+    public static bool WithChance(this Random randomizer, decimal chance) {
+        if (chance is < 0m or > 1m)
             throw new ArgumentOutOfRangeException($"{chance} must be between 0 and 1");
 
-        return chance > randomizer.NextDouble();
+        return chance > (decimal)randomizer.NextDouble();
     }
 }
